requirements.txt
# FastAPI et WebSocket
fastapi
uvicorn[standard]
websockets

# IA et Classification
sentence-transformers
scikit-learn
numpy
aiohttp
requests

# TTS
edge-tts

# Base de données
psycopg2-binary
asyncpg
redis

# Utilitaires
python-dotenv

